Safari-609.1.11.1/LayoutTests/accessibility/mac/aria-menuitem-focus-expected.txt

Item 0
Item 1
This tests that ARIA menuitems expose AXFocused correctly.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS item.isAttributeSupported('AXFocused') is true
PASS item.isAttributeSettable('AXFocused') is true
PASS item.boolAttributeValue('AXFocused') is false
PASS item.boolAttributeValue('AXFocused') is true
PASS item.isAttributeSupported('AXFocused') is true
PASS item.isAttributeSettable('AXFocused') is true
PASS item.boolAttributeValue('AXFocused') is false
PASS item.boolAttributeValue('AXFocused') is true
PASS successfullyParsed is true

TEST COMPLETE
